// src/app/model/university.model.ts
import { Grade } from '../interfaces/grade.interface';


export class University {
  id: string;
  name: string;
  info: string;
  scores: Grade;
  // id1: string;
  // nam2e: string;
  // info3: string;
  // id4: string;
  // nam5e: string;
  // info6?: boolean;

  // constructor(id: string, name: string, info: string, scores: Array<any>) {

  // }


  constructor(id: string, name: string, info: string, score: Grade) {
    // data =undefined;  undefined.id
    // data = data || {};
    // this.id = data.id;
    // this.info6 = data.info || false;
    // this.scores = data.scores || [];
  }

  // const newA: A = {name: 'asdasd'};
}

// const newUni = new University('NAM');

// const ojb = {
//   id: 'ABC',
//   name: 'adasd',
// };
// const newUni1 = new University(ojb);
// const newUni2 = new University();


// export interface A {
//   id: string;
//   age: number;


// }